Yolo-merge docker dependency updates (#152)
Version-pinned docker dependencies tend to break the build very very often. This gums up the rest of the renovate PRs and spends a lot of developer time.

I've never seen one of these build successfully and then break something, so I think it is safe enough to send these to the test federation if the build passes. Might need to revisit if we're ever considering deploying straight to production. 🤷