Tweak naming
Mostly change the name of the workflow to "Self-approve tidying" to make
it clearer what it's for. We don't want it to be used for non-tidying
changes.

Also change the command to /tidy for similar reasons. As a bonus, it's
easier to type.

Finally, change the names of some of the workflow steps. This lets us
remove an obvious comment.