Prepare to issue tidying licenses (#285)
* Tweak naming

Mostly change the name of the workflow to "Self-approve tidying" to make
it clearer what it's for. We don't want it to be used for non-tidying
changes.

Auto does kind of mean "self", but it more commonly means "automatic".
Self only means self, so it gives a better idea of what we actually
mean: these are the people who can approve their own PRs.

Also change the command to /tidy for similar reasons. As a bonus, it's
easier to type.

Finally, change the names of some of the workflow steps. This lets us
remove an obvious comment.

* Add CODEOWNERS for tidying licenses

If branch protections are configured appropriately, this will require
Russell or Matthew to approve changes to who can self-approve tidying
PRs.

In case of emergency (like me and Russell dying in a plane crash), it's
still possible for organisation admins to bypass these checks and merge
changes directly.

* Remove all tidying licenses

We'll add these back PR by PR.

* Don't delete /tidy comment

Probably want to keep these by default so that it's obvious why
aaf-terraform chose to approve the PR.