Fail finish step when intermediate step cancelled
Currently, if you let the init step of a deployment workflow complete,
then cancel further steps, you'll get a result string something like

successcancelledcancelledcancelledcancelledcancelledcancelledcancelledcancelledcancelledcancelledcancelledcancelledcancelled

Because this doesn't contain the substring "failed", the workflow is
incorrectly marked as finished and merging is allowed.

To fix this, grep for "cancelled" or "failed" in the result string, not
just "failed".