Reduce VS Code extension recommendations (#650)
Reasons applicable to ALL removals are:

- overhead from installing/patching/deprecating
- overwhelm for new devs.

More specifically:

* Don't copilot extensions

Comes packaged with VS code these days.

* Don't recommend gitignore extension

VS code understands .gitignore out of the box, and we don't need any
other features of the extension.

* Don't recommend icons extension

IMO, costs outweigh benefits:

Benefit:
- Icon for zillions of file types, looks nicer.

Purported benefit, but I don't care:
- Can bring own icons.

Cost:
- Makes it unclear whether you have language support set up for a file!!

* Don't recommend dotenv extension

The syntax highlighting of this extension implies that:

VAR_1=true
VAR_2=false
VAR_3=123

will all be different types when parsed. This is misleading: they'll all
be loaded as strings.

VS Code's built-in highlighting of .env correctly highlights all these
things in the same way.

* Remove reference to archived extension

* Don't recommend GitLens

This extension is quite slow.

Re: tooling, I see the best git results from command-line users. lazygit
also seems good. And VS Code's built-in GUI is at least okay.

GitLens doesn't seem like enough of an improvement over those (if it is
an improvement at all) to warrant a recommendation.

* Don't recommend markdown lint

Great tool, but can conflict with the formatting we have set up
elsewhere. (We format MD with prettier) If we're going to recommend an
extension across all projects, I want it to work immediately with 0
config, or at least do nothing.

* Don't recommend YAML extension

- Out-of-the-box support for YAML is good.
- The extension's formatter can conflict with what we have set up via
  prettier and ESLint.
- I'd rather get K8s support from a purpose-built extension.
- A 3.5 star rating is suspicious.

* Don't recommend shell-format

- Screws up Dockerfiles by default.
- Responsibility overlaps with other formatting tools we use.
- 3.5 star rating is suspicious.

* Remove duplicate recommendation

* Don't recommend stylelint

Out-of-the-box, this conflicts with VS Code's built-ins, causing bad DX.

We also don't incorporate this into CI anywhere, which is a recipe for
spurious diffs.

* Don't recommend Auto Import

- Hogs CPU.
- Typescript-only
  - doesn't apply to most projects.
  - VS Code's built-in typescript support is already excellent.

* Don't recommend Auto Rename Tag

- Corrupts undo stack.
- VS Code now has a built-in Linked Editing setting which does this with
  no bugs.

* Don't recommend Auto Close Tag

- Big CPU hog.
- Bug: sees some things as tags which aren't, e.g. in typescript
  List<string> will be auto-completed to List<string></string>, which is
  nonsense.

* Don't recommend niche tools

Totally fine, but too niche to be a broad recommendation. I don't want
to clog people's menus as they generally have an overwhelming amount of
items as-is.

* Don't recommend TODO highlight

The Better Comments extension already highlights todos.