Recommend fewer VS code extensions (#712)
* Don't recommend jinja or jest extensions

Most devs don't work with these techs often/

* Don't recommend spell checker extension

Not much of a value add in this age of LLMs.

* Don't recommend rdbg extension

Not necessary for most (all?) dev work. 99% (100%?) of Ruby dev work
can be debugged without this extension.

* Don't recommend babel extension

Built-in syntax highlighting is sufficient for us. We don't use graphql
or jsx, so we hardly need to have graphql highlighted within jsx.

* Don't recommend gutter preview extension

Not valuable often enough to be worth the bloat.